+++ Applications/Android/Samples/src/main/java/org/mapsforge/applications/android/samples/RotateMapViewer.java
@@ -30,18 +30,7 @@ import android.widget.Button;
 public class RotateMapViewer extends OverlayMapViewer {
 
 	@Override
-	protected void createMapViews() {
-		mapView = getMapView();
-		mapView.getModel().frameBufferModel.setOverdrawFactor(1.0d);
-		// mapView.getModel().init(this.preferencesFacade);
-		mapView.setClickable(true);
-		mapView.getMapScaleBar().setVisible(false);
-		mapView.setBuiltInZoomControls(hasZoomControls());
-		mapView.getMapZoomControls().setZoomLevelMin(getZoomLevelMin());
-		mapView.getMapZoomControls().setZoomLevelMax(getZoomLevelMax());
-		initializePosition(mapView.getModel().mapViewPosition);
-
-		// Rotate button
+	protected void createControls() {
 		Button rotateButton = (Button) findViewById(R.id.rotateButton);
 		rotateButton.setOnClickListener(new View.OnClickListener() {
 			@Override
@@ -53,14 +42,23 @@ public class RotateMapViewer extends OverlayMapViewer {
 		});
 	}
 
+	@Override
+	protected void createMapViews() {
+		mapView = getMapView();
+		mapView.getModel().frameBufferModel.setOverdrawFactor(1.0d);
+		// mapView.getModel().init(this.preferencesFacade);
+		mapView.setClickable(true);
+		mapView.getMapScaleBar().setVisible(false);
+		mapView.setBuiltInZoomControls(hasZoomControls());
+		mapView.getMapZoomControls().setZoomLevelMin(getZoomLevelMin());
+		mapView.getMapZoomControls().setZoomLevelMax(getZoomLevelMax());
+		initializePosition(mapView.getModel().mapViewPosition);
+	}
+
 	@TargetApi(Build.VERSION_CODES.HONEYCOMB_MR2)
 	@SuppressWarnings("deprecation")
 	@Override
 	protected void createTileCaches() {
-		boolean threaded = sharedPreferences.getBoolean(
-				SamplesApplication.SETTING_TILECACHE_THREADING, true);
-		int queueSize = Integer.parseInt(sharedPreferences.getString(
-				SamplesApplication.SETTING_TILECACHE_QUEUESIZE, "4"));
 		boolean persistent = sharedPreferences.getBoolean(
 				SamplesApplication.SETTING_TILECACHE_PERSISTENCE, true);
 
@@ -79,8 +77,7 @@ public class RotateMapViewer extends OverlayMapViewer {
 				getPersistableId(),
 				this.mapView.getModel().displayModel.getTileSize(), hypot,
 				hypot,
-				this.mapView.getModel().frameBufferModel.getOverdrawFactor(),
-				threaded, queueSize, persistent));
+				this.mapView.getModel().frameBufferModel.getOverdrawFactor(), persistent));
 	}
 
 	@Override