+++ mapsforge-core/src/main/java/org/mapsforge/core/mapelements/SymbolContainer.java
@@ -1,6 +1,7 @@
 /*
  * Copyright 2010, 2011, 2012, 2013 mapsforge.org
  * Copyright 2014-2015 Ludwig M Brinckmann
+ * Copyright 2015 devemux86
  *
  * This program is free software: you can redistribute it and/or modify it under the
  * terms of the GNU Lesser General Public License as published by the Free Software
@@ -25,8 +26,8 @@ import org.mapsforge.core.model.Rotation;
 
 public class SymbolContainer extends MapElementContainer {
 
-	private final boolean alignCanvas; // if it has a fixed angle to canvas.
-	private final boolean alignCenter;
+        final boolean alignCanvas; // if it has a fixed angle to canvas.
+        final boolean alignCenter;
 	public Bitmap symbol;
 	public final float theta;
 
@@ -76,7 +77,8 @@ public class SymbolContainer extends MapElementContainer {
 
 	public void draw(Canvas canvas, Point origin, Matrix matrix, final Rotation rotation) {
 		matrix.reset();
-		matrix.translate((float) (this.xy.x - origin.x + boundary.left), (float) (this.xy.y - origin.y + boundary.top));
+		// We cast to int for pixel perfect positioning
+		matrix.translate((int) (this.xy.x - origin.x + boundary.left), (int) (this.xy.y - origin.y + boundary.top));
 
 		float totalTheta = theta; // this is the rotation angle combined from map rotation and symbol rotation
 		if (!Rotation.noRotation(rotation) && this.alignCanvas) {