+++ mapsforge-map-android/src/main/java/org/mapsforge/map/android/graphics/AndroidCanvas.java
@@ -82,7 +82,6 @@ class AndroidCanvas implements Canvas {
 		if (paint.isTransparent()) {
 			return;
 		}
-
 		this.canvas.drawPath(AndroidGraphicFactory.getPath(path), AndroidGraphicFactory.getPaint(paint));
 	}