+++ mapsforge-map/src/main/java/org/mapsforge/map/layer/cache/FileSystemTileCache.java
@@ -1,8 +1,8 @@
 /*
  * Copyright 2010, 2011, 2012, 2013 mapsforge.org
- * Copyright 2014 Ludwig M Brinckmann
+ * Copyright 2014-2015 Ludwig M Brinckmann
  * Copyright 2014 mvglasow <michael -at- vonglasow.com>
- * Copyright 2014 devemux86
+ * Copyright 2014, 2015 devemux86
  *
  * This program is free software: you can redistribute it and/or modify it under the
  * terms of the GNU Lesser General Public License as published by the Free Software
@@ -25,8 +25,6 @@ import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.HashSet;
 import java.util.Set;
-import java.util.concurrent.LinkedBlockingQueue;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.locks.ReentrantReadWriteLock;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -36,54 +34,23 @@ import org.mapsforge.core.graphics.GraphicFactory;
 import org.mapsforge.core.graphics.TileBitmap;
 import org.mapsforge.core.util.IOUtils;
 import org.mapsforge.map.layer.queue.Job;
-import org.mapsforge.map.util.PausableThread;
-
-/**
- * Container class to tie a key/bitmap together.
- */
-class StorageJob {
-	Job key;
-	TileBitmap bitmap;
-
-	StorageJob(Job key, TileBitmap bitmap) {
-		this.key = key;
-		this.bitmap = bitmap;
-	}
-
-	/**
-	 * Equality is just defined over the key, not over the bitmap content. This allows
-	 * finding a StorageJob in the queue without knowing the data.
-	 */
-	@Override
-	public boolean equals(Object obj) {
-		if (this == obj) {
-			return true;
-		} else if (!(obj instanceof StorageJob)) {
-			return false;
-		}
-		StorageJob other = (StorageJob) obj;
-		return key.equals(other.key);
-	}
-
-	@Override
-	public int hashCode() {
-		return this.key.hashCode();
-	}
-}
+import org.mapsforge.map.model.common.Observable;
+import org.mapsforge.map.model.common.Observer;
 
 /**
  * A thread-safe cache for image files with a fixed size and LRU policy.
  * <p>
- * A {@code FileSystemTileCache} caches tiles in a dedicated path in the file system, specified in the constructor. The
- * cache writes the data on a separate thread, i.e. when the call to put a job/tile into the cache returns the data is
- * not actually written to disk.
+ * A {@code FileSystemTileCache} caches tiles in a dedicated path in the file system, specified in the constructor.
  * <p>
  * When used for a {@link org.mapsforge.map.layer.renderer.TileRendererLayer}, persistent caching may result in clipped
  * labels when tiles from different instances are used. To work around this, either display labels in a separate
  * {@link org.mapsforge.map.layer.labels.LabelLayer} (experimental) or disable persistence as described in
- * {@link #FileSystemTileCache(int, File, GraphicFactory, boolean, int, boolean)}.
+ * {@link #FileSystemTileCache(int, File, GraphicFactory, boolean)}.
+ *
+ * Note: previously the FileSystemTileCache utilized threading to speed up response times. This is not the
+ * case anymore and the constructors have been removed.
  */
-public class FileSystemTileCache extends PausableThread implements TileCache {
+public class FileSystemTileCache implements TileCache {
 	static final String FILE_EXTENSION = ".tile";
 	private static final Logger LOGGER = Logger.getLogger(FileSystemTileCache.class.getName());
 
@@ -94,7 +61,7 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 	 * of subdir of the cache dir (as in the standard TMS directory layout of zoomlevel/x/y). The relative path to the
 	 * cached tile, after stripping the extension, is used as the lookup key.
 	 */
-	class CacheDirectoryReader implements Runnable {
+	private class CacheDirectoryReader implements Runnable {
 		public void run() {
 			File[] zFiles = FileSystemTileCache.this.cacheDirectory.listFiles();
 			if (zFiles != null) {
@@ -136,11 +103,8 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 	 *            The File instance to examine. This can be null, which will cause the method to return {@code false}.
 	 */
 	private static boolean isValidCacheDirectory(File file) {
-		if ((file == null) || (!file.exists() && !file.mkdirs()) || !file.isDirectory() || !file.canRead()
-				|| !file.canWrite()) {
-			return false;
-		}
-		return true;
+		return !((file == null) || (!file.exists() && !file.mkdirs()) || !file.isDirectory() || !file.canRead()
+				|| !file.canWrite());
 	}
 
 	/**
@@ -167,6 +131,10 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 	 */
 
 	private static boolean deleteDirectory(File dir) {
+		if (dir == null) {
+			return false;
+		}
+
 		if (dir.isDirectory()) {
 			String[] children = dir.list();
 			if (children != null) {
@@ -184,15 +152,10 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 
 	private final File cacheDirectory;
 	private final GraphicFactory graphicFactory;
-	private final AtomicInteger jobs;
 	private FileWorkingSetCache<String> lruCache;
 	private final ReentrantReadWriteLock lock;
-	private boolean persistent;
-
-	// if threaded is true, the bitmap writing is executed on a separate thread,
-	// and jobs are stored in the jobStack. The false option remains for testing.
-	private final boolean threaded;
-	private final LinkedBlockingQueue<StorageJob> storageJobs;
+	private final Observable observable;
+	private final boolean persistent;
 
 	/**
 	 * Compatibility constructor that creates a non-threaded, non-persistent FSTC.
@@ -207,28 +170,7 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 	 *             if the capacity is negative.
 	 */
 	public FileSystemTileCache(int capacity, File cacheDirectory, GraphicFactory graphicFactory) {
-		this(capacity, cacheDirectory, graphicFactory, false, 0, false);
-	}
-
-	/**
-	 * Compatibility constructor that creates a non-persistent FSTC.
-	 * 
-	 * @param capacity
-	 *            the maximum number of entries in this cache.
-	 * @param cacheDirectory
-	 *            the directory where cached tiles will be stored.
-	 * @param graphicFactory
-	 *            the graphicFactory implementation to use.
-	 * @param threaded
-	 *            if cache will use background thread to store data (more responsive).
-	 * @param queueSize
-	 *            maximum length of queue before the put operation blocks
-	 * @throws IllegalArgumentException
-	 *             if the capacity is negative.
-	 */
-	public FileSystemTileCache(int capacity, File cacheDirectory, GraphicFactory graphicFactory, boolean threaded,
-			int queueSize) {
-		this(capacity, cacheDirectory, graphicFactory, threaded, queueSize, false);
+		this(capacity, cacheDirectory, graphicFactory, false);
 	}
 
 	/**
@@ -239,17 +181,13 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 	 * persistent cache will not delete the cache directory. Conversely, a non-persistent cache will serve only tiles
 	 * added to it via the {@link #put(Job, TileBitmap)} method, and calling {@link #destroy()} on a non-persistent
 	 * cache will delete {@code cacheDirectory}.
-	 * 
+	 *
 	 * @param capacity
 	 *            the maximum number of entries in this cache.
 	 * @param cacheDirectory
 	 *            the directory where cached tiles will be stored.
 	 * @param graphicFactory
 	 *            the graphicFactory implementation to use.
-	 * @param threaded
-	 *            if cache will use background thread to store data (more responsive).
-	 * @param queueSize
-	 *            maximum length of queue before the put operation blocks
 	 * @param persistent
 	 *            if cache data will be kept between instances
 	 * @throws IllegalArgumentException
@@ -257,16 +195,9 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 	 * @throws IllegalArgumentException
 	 *             if the capacity is negative.
 	 */
-	public FileSystemTileCache(int capacity, File cacheDirectory, GraphicFactory graphicFactory, boolean threaded,
-			int queueSize, boolean persistent) {
-		this.jobs = new AtomicInteger(0);
+	public FileSystemTileCache(int capacity, File cacheDirectory, GraphicFactory graphicFactory, boolean persistent) {
+		this.observable = new Observable();
 		this.persistent = persistent;
-		this.threaded = threaded;
-		if (threaded) {
-			this.storageJobs = new LinkedBlockingQueue<>(queueSize);
-		} else {
-			this.storageJobs = null;
-		}
 		this.lruCache = new FileWorkingSetCache<>(capacity);
 		this.lock = new ReentrantReadWriteLock();
 		if (isValidCacheDirectory(cacheDirectory)) {
@@ -282,9 +213,6 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 			this.cacheDirectory = null;
 		}
 		this.graphicFactory = graphicFactory;
-		if (this.threaded) {
-			this.start();
-		}
 	}
 
 	@Override
@@ -293,7 +221,7 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 			lock.readLock().lock();
 			// if we are using a threaded cache we return true if the tile is still in the
 			// queue to reduce double rendering
-			return (this.lruCache.containsKey(key.getKey()) || (threaded && storageJobs.contains(key)));
+			return this.lruCache.containsKey(key.getKey());
 		} finally {
 			lock.readLock().unlock();
 		}
@@ -307,10 +235,10 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 	 * If the cache is not persistent, calling this method is equivalent to calling {@link #purge()}. If the cache is
 	 * persistent, it does nothing.
 	 * <p>
-	 * Beginning with 0.6.0, accessing the cache after calling {@code destroy()} is discouraged. In order to empty the
+	 * Beginning with 0.5.1, accessing the cache after calling {@code destroy()} is discouraged. In order to empty the
 	 * cache and force all tiles to be re-rendered or re-requested from the source, use {@link #purge()} instead.
 	 * Earlier versions lacked the {@link #purge()} method and used {@code destroy()} instead, but this practice is now
-	 * discouraged and may lead to unexpected results when used with features introduced in 0.6.0 or later.
+	 * discouraged and may lead to unexpected results when used with features introduced in 0.5.1 or later.
 	 */
 	@Override
 	public void destroy() {
@@ -347,7 +275,7 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 			return null;
 		} catch (IOException e) {
 			remove(key);
-			LOGGER.log(Level.SEVERE, null, e);
+			LOGGER.log(Level.SEVERE, e.getMessage(), e);
 			return null;
 		} finally {
 			IOUtils.closeQuietly(inputStream);
@@ -391,30 +319,18 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 	 * rendered tiles, or the source for downloaded tiles. Applications which frequently alternate between a limited
 	 * number of map model configurations may want to consider using a different cache for each.
 	 * 
-	 * @since 0.6.0
+	 * @since 0.5.1
 	 */
 	@Override
 	public void purge() {
 		try {
 			this.lock.writeLock().lock();
 			this.lruCache.clear();
-			if (this.threaded) {
-				this.interrupt();
-			}
 		} finally {
 			this.lock.writeLock().unlock();
 		}
 
 		deleteDirectory(this.cacheDirectory);
-	}
-
-	/**
-	 * Gets the number of remaining tiles still in the queue to be written to disk.
-	 * 
-	 * @return number of jobs in queue, 0 if not threaded.
-	 */
-	public int getQueueLength() {
-		return jobs.get();
 	}
 
 	@Override
@@ -429,13 +345,8 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 			return;
 		}
 
-		jobs.incrementAndGet();
-		if (this.threaded) {
-			bitmap.incrementRefCount();
-			storageJobs.offer(new StorageJob(key, bitmap));
-		} else {
-			storeData(key, bitmap);
-		}
+		storeData(key, bitmap);
+		this.observable.notifyObservers();
 	}
 
 	public void setWorkingSet(Set<Job> workingSet) {
@@ -445,6 +356,17 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 		}
 		this.lruCache.setWorkingSet(workingSetInteger);
 	}
+
+	@Override
+	public void addObserver(final Observer observer) {
+		this.observable.addObserver(observer);
+	}
+
+	@Override
+	public void removeObserver(final Observer observer) {
+		this.observable.removeObserver(observer);
+	}
+
 
 	private File getOutputFile(Job job) {
 		String file = this.cacheDirectory + File.separator + job.getKey();
@@ -463,25 +385,6 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 			lock.writeLock().unlock();
 		}
 
-	}
-
-	protected void doWork() throws InterruptedException {
-		StorageJob x = storageJobs.take();
-		storeData(x.key, x.bitmap);
-	}
-
-	/**
-	 * @return the priority which will be set for this thread.
-	 */
-	protected ThreadPriority getThreadPriority() {
-		return ThreadPriority.BELOW_NORMAL;
-	}
-
-	/**
-	 * @return true if this thread has some work to do, false otherwise.
-	 */
-	protected boolean hasWork() {
-		return true;
 	}
 
 	/**
@@ -528,10 +431,6 @@ public class FileSystemTileCache extends PausableThread implements TileCache {
 			}
 		} finally {
 			IOUtils.closeQuietly(outputStream);
-			if (threaded) {
-				bitmap.decrementRefCount();
-			}
-			jobs.decrementAndGet();
 		}
 
 	}