+++ mapsforge-map/src/main/java/org/mapsforge/map/layer/cache/TileCache.java
@@ -19,11 +19,13 @@ import java.util.Set;
 
 import org.mapsforge.core.graphics.TileBitmap;
 import org.mapsforge.map.layer.queue.Job;
+import org.mapsforge.map.model.common.ObservableInterface;
+import org.mapsforge.map.model.common.Observer;
 
 /**
  * Interface for tile image caches.
  */
-public interface TileCache {
+public interface TileCache extends ObservableInterface {
 
 	/**
 	 * @return true if this cache contains an image for the given key, false otherwise.
@@ -36,10 +38,10 @@ public interface TileCache {
 	 * <p>
 	 * Applications are expected to call this method when they no longer require the cache.
 	 * <p>
-	 * In versions prior to 0.6.0, it was common practice to call this method but continue using the cache, in order to
-	 * empty it, forcing all tiles to be re-rendered or re-requested from the source. Beginning with 0.6.0,
+	 * In versions prior to 0.5.1, it was common practice to call this method but continue using the cache, in order to
+	 * empty it, forcing all tiles to be re-rendered or re-requested from the source. Beginning with 0.5.1,
 	 * {@link #purge()} should be used for this purpose. The earlier practice is now discouraged and may lead to
-	 * unexpected results when used with features introduced in 0.6.0 or later.
+	 * unexpected results when used with features introduced in 0.5.1 or later.
 	 */
 	void destroy();
 
@@ -75,7 +77,7 @@ public interface TileCache {
 	 * rendered tiles, or the source for downloaded tiles. Applications which frequently alternate between a limited
 	 * number of map model configurations may want to consider using a different cache for each.
 	 * 
-	 * @since 0.6.0
+	 * @since 0.5.1
 	 */
 	void purge();